requirements.txt
python-dotenv==0.20.0
pymongo==4.1.1
APScheduler==3.9.1
metaapi-cloud-copyfactory-sdk==7.1.1
metaapi-cloud-metastats-sdk==3.3.1
metaapi-cloud-sdk==21.5.2